+++ conan/api/subapi/upload.py
@@ -1,11 +1,8 @@
 from conan.api.output import ConanOutput
 from conan.api.subapi import api_method
 from conan.internal.conan_app import ConanApp
-from conan.internal.api.select_pattern import SelectPattern
-from conans.client.cmd.uploader import IntegrityChecker, PackagePreparator, UploadExecutor, \
-    UploadUpstreamChecker
+from conans.client.cmd.uploader import PackagePreparator, UploadExecutor, UploadUpstreamChecker
 from conans.client.pkg_sign import PkgSignaturesPlugin
-from conans.errors import ConanException
 
 
 class UploadAPI:
@@ -14,18 +11,11 @@ class UploadAPI:
         self.conan_api = conan_api
 
     @api_method
-    def get_bundle(self, expression, package_query=None, only_recipe=False):
-        ref_pattern = SelectPattern(expression)
-        if only_recipe:
-            if ref_pattern.package_id:
-                raise ConanException("Do not specify 'package_id' with 'only-recipe'")
-        else:
-            ref_pattern.package_id = ref_pattern.package_id or "*"
-        select_bundle = self.conan_api.list.select(ref_pattern, package_query)
-
-        # This is necessary to upload_policy = "skip"
+    def check_upstream(self, package_list, remote, force=False):
+        """Check if the artifacts are already in the specified remote, skipping them from
+        the package_list in that case"""
         app = ConanApp(self.conan_api.cache_folder)
-        for ref, bundle in select_bundle.refs():
+        for ref, bundle in package_list.refs():
             layout = app.cache.ref_layout(ref)
             conanfile_path = layout.conanfile()
             conanfile = app.loader.load_basic(conanfile_path)
@@ -34,37 +24,23 @@ class UploadAPI:
                                    "because upload_policy='skip'")
                 bundle["packages"] = {}
 
-        return select_bundle
+        UploadUpstreamChecker(app).check(package_list, remote, force)
 
     @api_method
-    def check_integrity(self, upload_data):
-        """Check if the recipes and packages are corrupted (it will raise a ConanExcepcion)"""
-        app = ConanApp(self.conan_api.cache_folder)
-        checker = IntegrityChecker(app)
-        checker.check(upload_data)
-
-    @api_method
-    def check_upstream(self, upload_bundle, remote, force=False):
-        """Check if the artifacts are already in the specified remote, skipping them from
-        the upload_bundle in that case"""
-        app = ConanApp(self.conan_api.cache_folder)
-        UploadUpstreamChecker(app).check(upload_bundle, remote, force)
-
-    @api_method
-    def prepare(self, upload_bundle, enabled_remotes):
+    def prepare(self, package_list, enabled_remotes):
         """Compress the recipes and packages and fill the upload_data objects
         with the complete information. It doesn't perform the upload nor checks upstream to see
         if the recipe is still there"""
         app = ConanApp(self.conan_api.cache_folder)
         preparator = PackagePreparator(app)
-        preparator.prepare(upload_bundle, enabled_remotes)
+        preparator.prepare(package_list, enabled_remotes)
         signer = PkgSignaturesPlugin(app.cache)
-        # This might add files entries to upload_bundle with signatures
-        signer.sign(upload_bundle)
+        # This might add files entries to package_list with signatures
+        signer.sign(package_list)
 
     @api_method
-    def upload_bundle(self, upload_bundle, remote):
+    def upload(self, package_list, remote):
         app = ConanApp(self.conan_api.cache_folder)
         app.remote_manager.check_credentials(remote)
         executor = UploadExecutor(app)
-        executor.upload(upload_bundle, remote)
+        executor.upload(package_list, remote)
